Add key generator container image
* adds key generator container image
* replace Volkan’s email with a generic maintainer’s email in the LABELs of the container images.
