+++ app/keygen/README.md
@@ -1,0 +1,18 @@
+```text
+|   Protect your secrets, protect your sensitive data.
+:   Explore VMware Secrets Manager docs at https://vsecm.com/
+</
+<>/ keep your secrets… secret
+```
+
+## VMware Secrets Manager (VSecM) Keygen
+
+**VSecM Keygen** is a utility application that generates VSecM Safe’s 
+bootstrapping keys if you want an extra level of security and control the
+creation of the master key yourself (*instead of letting VSecM Safe generate
+it for you*).
+
+This is useful for **manual** operation mode of **VSecM Safe** where it will
+remain locked until you manually unlock it by providing the necessary keys.
+
+Check out more details in the [official documentation](https://vsecm.com/).
