+++ dockerfiles/vsecm-photon-fips/sidecar.Dockerfile
@@ -24,7 +24,7 @@ RUN CGO_ENABLED=0 GOEXPERIMENT=boringcrypto GOOS=linux go build -mod vendor -a -
 # generate clean, final image for end users
 FROM photon:5.0
 
-LABEL "maintainers"="Volkan Özçelik <volkan@vsecm.com>"
+LABEL "maintainers"="VSecM Maintainers <maintainers@vsecm.com>"
 LABEL "version"="0.21.0"
 LABEL "website"="https://vsecm.com/"
 LABEL "repo"="https://github.com/vmware-tanzu/secrets-manager"
